Refactor: move TopicForm into AddTopic
